Update Readme Description
A simple hello world in Electron plus some other basics and experimentation. Nothing to show off, of course, but I'll make it public because I like to share what I'm learning (or how I started learning, it depends when you are reading this) with my friends following me and/or any other stalker. - March 11, 2018.